# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Formula 1 Mock API
  version: "1.0.0"
  description: Mock API based on Ergast Developer API structure for local development

servers:
  - url: http://localhost:3000/api

paths:
  /seasons:
    get:
      summary: Get all seasons
      responses:
        "200":
          description: List of F1 seasons

  /seasons/{season}:
    get:
      summary: Get details for a specific season
      parameters:
        - name: season
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Season detail
        "404":
          description: Season not found

  /seasons/{season}/races:
    get:
      summary: Get all races for a season
      parameters:
        - name: season
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of races

  /{season}/drivers:
    get:
      summary: Get all drivers for a season
      parameters:
        - name: season
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of drivers

  /{season}/constructors:
    get:
      summary: Get all constructors for a season
      parameters:
        - name: season
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of constructors

  /{season}/{round}/results:
    get:
      summary: Get race results for a specific round
      parameters:
        - name: season
          in: path
          required: true
          schema:
            type: string
        - name: round
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Race results

  /{season}/driver-standings:
    get:
      summary: Get driver standings
      parameters:
        - name: season
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Driver standings

  /{season}/constructor-standings:
    get:
      summary: Get constructor standings
      parameters:
        - name: season
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Constructor standings

  /{season}/{round}/qualifying:
    get:
      summary: Get qualifying results for a round
      parameters:
        - name: season
          in: path
          required: true
          schema:
            type: string
        - name: round
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Qualifying results

  /{season}/{round}/pitstops:
    get:
      summary: Get pit stop data for a round
      parameters:
        - name: season
          in: path
          required: true
          schema:
            type: string
        - name: round
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pit stop data

  /{season}/{round}/laps:
    get:
      summary: Get lap times for a round
      parameters:
        - name: season
          in: path
          required: true
          schema:
            type: string
        - name: round
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lap times

  /status:
    get:
      summary: Get list of status codes
      responses:
        "200":
          description: Status info